Changes Andy wanted
Legacy video to be at the top of the About page instead, easy fix.

And let's change the font to match the other video fonts, no line graphic, just the font. And let's try Your Labels, Our Story on the same line.

Also, let's change the header tab and header on the top of the page that says 'PRINTING' to 'CAPABILITIES'